feat: Porting the latest known simulation code to this repository
NOTE: I have moved all the old simulation scripts to a folder called Old Simulator (and prepended "Old" to some of the script names to avoid a duplicate script name error), and I have similarly renamed the old simulation scene to Old Simulator. This means that the old simulator scene WILL BE BROKEN given that the scripts it references are no longer in the place it expects. I don't imagine this will be an issue since we will be building off of the new simulator, but if it is then future code maintainers know why it happened.