Porting the latest known simulation code to this repository (again)
We got the latest code!
Deleting the OldSim code from the last commit because we can just go back to a previous milestone branch if we want to look at that again.

Also cleaned up the organization on some of the game objects, renaming objects to have more appropriate names as to what they're actually for, and updated the placeholder text in the test UI to use Azimuth and Elevation language instead of MoveZ and MoveY.