housekeeping: Clean up and add comments to the newest scripts
Also deleted some unused scripts.

Skipped over the SimServer script given that it's massive and we're probably better off going through it as a team.

Also only did cleaning of TelescopeControllerSim, formatting the file to the same style as the rest of the scripts and replacing some code with functionally equivalent code that looks cleaner. I'm pretty sure that this script is the source of #85, so we should step through this to see where that is arising.

Closes #108.