fix: TelescopeControllerSim ignored all new commands instead of treating azimuth and elevation commands seperately
Previously if the azimuth was moving and you sent a command to change the azimuth and elevation at the same time, it would ignore the azimuth changes but accept the elevation changes. Now no new commands can be received while a command is being executed.

Need to investigate if this is intended behavior for the real telescope.