fix: Convert speed from steps to degrees
Looks like the reason the received speed was so fast is because we were never converting it from steps/min to degrees/min.

Making this change broke jogging such that you could only ever jog up or to the right no matter which direction you chose. Will need to investigate this further in the future. It looks like part of the issue is that jogging can result in negative speed, but custom orientation movements never send negative speeds, or not so far as I had seen.

Fixes #130
Fixes #120